Support high resolution H264 (simulated YUV444) and make it default
Encode H264 frames with 200% width and height so that the encoded frames contains YUV444 equivalent colour information. With this approach the CPU and networking consumption remains at the same level as H264 encoding under the original resolution, but the image quality becomes at good as VP9 with YUV444 colour space.